Add minimal imposter based mock service handler
Imposter (https://github.com/outofcoffee/imposter) based simulator to test
wheather this kind of mock services can be used as a more lightweight
alternative to running the real microservices during development.

The system can be started from the simposter.sh script which will start
imposter instances in docker containers.

This implementation lacks QoL features (such as a way to stop the services)
and is intended mainly for getting feedback for further development.